pyproject.toml
[tool.poetry]
name = "FinalProject"
version = "0.1.0"
description = ""
authors = ["PotemkinaV"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pillow = "^10.1.0"
telebot = "^0.0.5"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"
scikit-learn = "^1.3.2"
tqdm = "^4.66.1"
torch = "^2.1.0"
ru-core-news-sm = {url = "https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0.tar.gz"}
sumy = "^0.11.0"
textract = "^1.6.5"
transformers = "^4.34.1"
streamlit = "^1.28.0"
openai = "^1.10.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
